Original programs uplooaded
Upload 2 Mathematica v12 programs.  The first displays images in the CheXpert dataset that I identified as damaged.  The second attempts to correct the damage to images produced by shift of rows of the image.  Not all damaged images can be repaired.  One  damaged image appears to be a double exposure.  Other images have blanked rows.  These cannot be corrected.  The majority of damaged images due to shifted rows can be corrected.